add roles service

+++ backend/src/main/java/com/example/backend/controller/RoleController.java
@@ -35,7 +35,7 @@ class RoleController {
     }
 
     @GetMapping("roles/{id}")
-    public ResponseEntity<ApiRespone<Role>> getRole(@PathVariable int id){
+    public ResponseEntity<ApiRespone<Role>> getRole(@PathVariable Long id){
         
         Role data = roleService.getRollById(id);
         if(data == null) return ResponseEntity.ok(new ApiRespone<>(false, null, "Roles Not Found"));
